+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Do not set executable bit for out-dated files (since i-doit 1.10.1)
 -   Cancel script if user likes to
+-   SLES: install PHP modules bz2, memcached and posix
 
 
 ##  [0.9] – 2017-12-19